refactor(background): move context-menu constants out of constants/
The `context-menu.ts` file held UI wiring constants consumed only by the
background service worker, not a shared catalogue. Move it next to
`background.ts` and drop the `constants/` directory that only hosted
this one file.
